# specs/003-add-a-matching/contracts/api.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  version: 1.0.0
  description: |
    API contracts for the admin dashboard content management system.
    Extends existing blog and portfolio APIs with admin-specific operations.
  contact:
    name: Rohan Batra
    email: contact@rohanbatra.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://rohanbatra.com/api
    description: Production server
security:
  - ClerkAuth: []
paths:
...
  /admin/stats:
    get:
      summary: Get dashboard statistics and overview data
      tags: [Admin Dashboard]
      security:
        - ClerkAuth: []
      responses:
        '200':
          description: Dashboard statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/users/{userId}:
    get:
      summary: Get specific user details
      tags: [Admin User Management]
      security:
        - ClerkAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetail'

    put:
      summary: Update user role or status
      tags: [Admin User Management]
      security:
        - ClerkAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'

  # Lottie Asset Management
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      description: Clerk JWT token in Authorization header
  schemas:
...
    DashboardStats:
      type: object
      properties:
        totalBlogPosts:
          type: integer
        totalProjects:
          type: integer
        totalComments:
          type: integer
        totalUsers:
          type: integer
        totalAssets:
          type: integer
        pendingComments:
          type: integer
        publishedPosts:
          type: integer
        featuredProjects:
          type: integer
        storageUsed:
          type: integer
          description: Total storage used in bytes
        topPerformingPosts:
          type: array
          items:
            type: object
            properties:
              slug:
                type: string
              title:
                type: string
              viewCount:
                type: integer
              likeCount:
                type: integer
...
    Comment:
      type: object
      properties:
        _id:
          type: string
        postSlug:
          type: string
        userId:
          type: string
        content:
          type: string
        parentId:
          type: string
        status:
          type: string
          enum: [pending, published, spam, deleted]
        language:
          type: string
        metadata:
          type: object
          properties:
            ipAddress:
              type: string
            userAgent:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminUser:
      type: object
      properties:
        _id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, editor, admin]
        avatar:
          type: string
        bio:
          type: string
        preferences:
          type: object
        activity:
          type: object
          properties:
            commentCount:
              type: integer
            likeCount:
              type: integer
            lastActive:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminUserDetail:
      allOf:
        - $ref: '#/components/schemas/AdminUser'
        - type: object
          properties:
            recentComments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
            recentLikes:
              type: array
              items:
                type: object
                properties:
                  postSlug:
                    type: string
                  postTitle:
                    type: string
                  type:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
    UpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [user, editor, admin]
        status:
          type: string
          enum: [active, suspended, banned]

    # Asset Admin Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...